+++ js/main.js
@@ -1,104 +1,3 @@
-function getRandomNumber(from, to) {
-  if (from < 0 || to < 0) {
-    return false;
-  }
-  if (typeof from !== 'number' || typeof to !== 'number') {
-    return false;
-  }
-  if (from === to) {
-    return from;
-  }
-  if (from > to) {
-    [from, to] = [to, from];
-  }
-  return Math.round(Math.random() * (to - from) + from);
-}
-function lineMaxLength(line, maxLength) {
-  return line.length < maxLength;
-}
-const DESCRIPTIONS = [
-  'hi',
-  'sunset',
-  'night',
-  'you',
-  'birff',
-  'stars',
-  'morning',
-  'happy!',
-  'a wish',
-  'moon',
-  'food',
-  'friends',
-  'love',
-  'what',
-  'who',
-  'ambition',
-  'envy',
-  'huh',
-  'spring',
-  'new year'
-];
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAMES = [
-  'Luna',
-  'Oscar',
-  'Aurora',
-  'Iris',
-  'Cora',
-  'Atlas',
-  'Freya',
-  'Finn',
-  'Jasper',
-  'Flora',
-  'Dylan',
-  'Orion',
-  'Ariadne',
-  'Astra',
-  'Remus'
-];
-const PHOTO_DESCRIPTIONS_NUMBER = 25;
-const getRandomArrayElement = (elements) => {
-  return elements[getRandomNumber(0, elements.length - 1)];
-};
-function createUniqueRandomId (from, to) {
-  const previousValues = [];
-  return function () {
-    let currentValue = getRandomNumber(from, to);
-    if (previousValues.length >= (max - min + 1)) {
-      return false
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomNumber(from, to);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  }
-}
-const createComment = () => {
-  return {
-    id: createUniqueRandomId(1, 600),
-    avatar: 'img/avatar-' + getRandomNumber(1, 6) + '.svg',
-    message: getRandomArrayElement(MESSAGES),
-    name: getRandomArrayElement(NAMES)
-  };
-};
-const createObject = () => {
-  return {
-    id: createUniqueRandomId(1, 25),
-    url: 'photos/' + createUniqueRandomId(1, 25),
-    description: getRandomArrayElement(DESCRIPTIONS),
-    likes: getRandomNumber(15, 200),
-    comments: Array.from({length: 3}, createComment)
-  };
-};
-const photo_descriptions = (number) => {
-  Array.from({length: number}, createObject);
-};
-console.log(photo_descriptions(PHOTO_DESCRIPTIONS_NUMBER));
+import {photoDescriptions} from './data.js';
+
+console.log(photoDescriptions);
